UserAccess property of Organization model is now moved into a separate extended class: UserOrganization.
